+++ packages/openapi-validation/CHANGELOG.md
@@ -1,5 +1,12 @@
+<a name="openapi-validation-2.7.0"></a>
+# [openapi-validation-2.7.0](https://github.com/ditsmod/ditsmod/releases/tag/openapi-validation-2.7.0) (2023-01-12)
+
+### BREAKING CHANGES
+
+- Migrate to core-2.35.0.
+
 <a name="openapi-validation-2.6.0"></a>
-# [openapi-validation-2.6.0](https://github.com/ditsmod/ditsmod/releases/tag/openapi-validation-2.6.0) (2023-01-07)
+## [openapi-validation-2.6.0](https://github.com/ditsmod/ditsmod/releases/tag/openapi-validation-2.6.0) (2023-01-07)
 
 ### BREAKING CHANGES